+++ chainstate/src/detail/mod.rs
@@ -31,15 +31,7 @@ use std::{collections::VecDeque, sync::Arc};
 
 use itertools::Itertools;
 use thiserror::Error;
-use utils_networking::broadcaster;
 
-use self::{
-    block_invalidation::BlockInvalidator,
-    orphan_blocks::{OrphanBlocksMut, OrphansProxy},
-    query::ChainstateQuery,
-    tx_verification_strategy::TransactionVerificationStrategy,
-};
-use crate::{BlockInvalidatorError, ChainstateConfig, ChainstateEvent};
 use chainstate_storage::{
     BlockchainStorage, BlockchainStorageRead, BlockchainStorageWrite, TransactionRw, Transactional,
 };
@@ -62,13 +54,26 @@ use pos_accounting::{
 use tx_verifier::transaction_verifier;
 use utils::{
     const_value::ConstValue,
-    ensure,
+    debug_assert_or_log, ensure,
     eventhandler::{EventHandler, EventsController},
     log_error,
     set_flag::SetFlag,
     tap_log::TapLog,
 };
+use utils_networking::broadcaster;
 use utxo::UtxosDB;
+
+use crate::{
+    detail::bootstrap::import_bootstrap_stream, BlockInvalidatorError, BootstrapError,
+    ChainstateConfig, ChainstateError, ChainstateEvent,
+};
+
+use self::{
+    block_invalidation::BlockInvalidator,
+    orphan_blocks::{OrphanBlocksMut, OrphansProxy},
+    query::ChainstateQuery,
+    tx_verification_strategy::TransactionVerificationStrategy,
+};
 
 pub use self::{
     error::*, info::ChainInfo, median_time::calculate_median_time_past,
@@ -176,9 +181,7 @@ impl<S: BlockchainStorage, V: TransactionVerificationStrategy> Chainstate<S, V> 
         tx_verification_strategy: V,
         custom_orphan_error_hook: Option<Arc<OrphanErrorHandler>>,
         time_getter: TimeGetter,
-    ) -> Result<Self, crate::ChainstateError> {
-        use crate::ChainstateError;
-
+    ) -> Result<Self, ChainstateError> {
         let best_block_id = {
             let db_tx = chainstate_storage
                 .transaction_ro()
@@ -204,6 +207,12 @@ impl<S: BlockchainStorage, V: TransactionVerificationStrategy> Chainstate<S, V> 
         }
 
         chainstate.update_initial_block_download_flag()?;
+
+        if !chainstate.is_initial_block_download_finished.test()
+            && chainstate.chainstate_config.db_reckless_mode_in_ibd_enabled()
+        {
+            chainstate.chainstate_storage.set_reckless_mode(true)?;
+        }
 
         chainstate
             .check_consistency()
@@ -780,6 +789,14 @@ impl<S: BlockchainStorage, V: TransactionVerificationStrategy> Chainstate<S, V> 
 
         if self.is_fresh_block(&tip_timestamp) {
             self.is_initial_block_download_finished.set();
+
+            if self.chainstate_storage.in_reckless_mode()? {
+                debug_assert_or_log!(
+                    self.chainstate_config.db_reckless_mode_in_ibd_enabled(),
+                    "The db was in a reckless mode even though it wasn't enabled"
+                );
+                self.chainstate_storage.set_reckless_mode(false)?;
+            }
         }
 
         Ok(())
@@ -816,6 +833,42 @@ impl<S: BlockchainStorage, V: TransactionVerificationStrategy> Chainstate<S, V> 
             Ok(_) => Ok(()),
             Err(err) => (*err).into(),
         }
+    }
+
+    #[log_error]
+    pub fn import_bootstrap_stream<'a>(
+        &mut self,
+        mut reader: std::io::BufReader<Box<dyn std::io::Read + Send + 'a>>,
+    ) -> Result<(), BootstrapError> {
+        let enable_reckless_mode = !self.chainstate_storage.in_reckless_mode()?
+            && self.chainstate_config.db_reckless_mode_in_ibd_enabled();
+
+        if enable_reckless_mode {
+            self.chainstate_storage.set_reckless_mode(true)?;
+        }
+
+        let chain_config = Arc::clone(&self.chain_config);
+        let mut block_processor = |block: WithId<Block>| -> Result<_, BootstrapError> {
+            let block_exists = self.make_db_tx_ro()?.block_exists(&block.get_id())?;
+
+            if !block_exists {
+                Ok(self.process_block(block, BlockSource::Local)?)
+            } else {
+                Ok(None)
+            }
+        };
+
+        let result = import_bootstrap_stream(&chain_config, &mut reader, &mut block_processor);
+
+        // FIXME write to log when reckless mode is turned on and off?
+
+        // FIXME check if the db can become corrupted silently
+
+        if enable_reckless_mode {
+            self.chainstate_storage.set_reckless_mode(false)?;
+        }
+
+        result
     }
 }