+++ node-lib/src/options.rs
@@ -239,6 +239,19 @@ pub struct RunOptions {
     #[clap(long, value_name = "COUNT")]
     pub max_db_commit_attempts: Option<usize>,
 
+    /// Whether to switch to the "reckless" mode during the initial block download or bootstrapping.
+    /// 
+    /// In the "reckless" mode the chainstate db contents is not synced to disk on each commit, which
+    /// increases performance at the cost of a potential db corruption if the system crashes.
+    /// 
+    /// Once the initial block download or bootstrapping is complete, the node will automatically
+    /// switch to the normal mode of operation.
+    /// 
+    /// Note: if a system crash does occur during syncing in the reckless mode and the chainstate
+    /// db gets corrupted, you may need to delete it manually and re-sync again.
+    #[clap(long, action = clap::ArgAction::SetTrue)]
+    pub enable_db_reckless_mode_in_ibd: Option<bool>,
+
     /// The maximum capacity of the orphan blocks pool in blocks.
     #[clap(long, value_name = "COUNT")]
     pub max_orphan_blocks: Option<usize>,
@@ -413,6 +426,7 @@ mod tests {
             node_type: Default::default(),
             mock_time: Default::default(),
             max_db_commit_attempts: Default::default(),
+            enable_db_reckless_mode_in_ibd: Default::default(),
             max_orphan_blocks: Default::default(),
             p2p_networking_enabled: Default::default(),
             p2p_bind_addresses: Default::default(),